PR7.1/weather_app/requirements.txt
Flask==3.0.3
gunicorn==22.0.0

SQLAlchemy==2.0.32
psycopg2-binary==2.9.9

Flask-Login==0.6.3
Flask-Session==0.8.0
redis==5.0.8

requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

python-dotenv==1.0.1
Werkzeug==3.0.3

python-dateutil==2.9.0.post0

Faker==26.0.0
matplotlib==3.9.2
pillow==10.4.0
pandas==2.2.2
